Separated codes in separate files by menu grouping
Also updated global variables and added number padding for readability.
