--- Includes/global.h
@@ -1,31 +1,0 @@
-#ifndef GLOBAL_H
-#define GLOBAL_H
-
-extern char currentEXP[40];
-extern char statusOutlines[40];
-extern char storedGender[40];
-extern char storedSkintone[40];
-extern int i_increaseEXP;
-extern int i_increaseEXP10;
-extern int i_decreaseEXP;
-extern int i_decreaseEXP10;
-extern int i_resetEXP;
-extern int i_walkThruWalls;
-extern int i_shinyPokemon;
-extern int i_instantEgg;
-extern int i_allItems;
-extern int i_allMedicine;
-extern int i_allTMs;
-extern int i_allBerries;
-extern int i_allClothes;
-extern int i_toggleOutlines;
-extern int i_rematchTrainers;
-extern int i_pcAnywhere;
-extern u8  currentGender;
-extern u8  currentSkintone;
-extern u8  matchingHair;
-extern int i_setGender;
-extern int i_setSkintone;
-extern int i_switchLooks;
-
-#endif
