+++ Sources/global.h
@@ -1,0 +1,33 @@
+#ifndef GLOBAL_H
+#define GLOBAL_H
+
+extern char currentEXP[40],
+            statusOutlines[40],
+            storedGender[40],
+            storedSkintone[40];
+
+extern int  i_increaseEXP,
+            i_increaseEXP10,
+            i_decreaseEXP,
+            i_decreaseEXP10,
+            i_resetEXP,
+            i_walkThruWalls,
+            i_shinyPokemon,
+            i_instantEgg,
+            i_allItems,
+            i_allMedicine,
+            i_allTMs,
+            i_allBerries,
+            i_allClothes,
+            i_toggleOutlines,
+            i_rematchTrainers,
+            i_pcAnywhere,
+            i_setGender,
+            i_setSkintone,
+            i_switchLooks;
+
+extern u8   currentGender,
+            currentSkintone;
+
+
+#endif
